pyproject.toml
[tool.poetry]
name = "django-demo"
version = "0.1.0"
description = ""
authors = ["Mark Nicholson <wmark.nicholsonu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.2.9"
python-dotenv = "^0.19.1"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
